fix(dogfood): normalize MCP counts and bound telemetry locks
