--- .github/workflows/apply-final-dogfood-repairs.yml
@@ -1,63 +1,0 @@
-name: Apply Final Dogfood Repairs
-
-on:
-  pull_request:
-    branches: [main]
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    if: >-
-      github.event.pull_request.head.ref == 'dogfood/deep-runtime-20260727' &&
-      github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    timeout-minutes: 25
-    steps:
-      - uses: actions/checkout@v6
-        with:
-          ref: dogfood/deep-runtime-20260727
-          fetch-depth: 0
-          token: ${{ secrets.GITHUB_TOKEN }}
-
-      - uses: actions/setup-python@v7
-        with:
-          python-version: '3.12'
-
-      - name: Apply exact-count replacements
-        run: python scripts/apply_final_dogfood_repairs.py
-
-      - name: Install focused test dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install -e '.[test,proxy]' pytest-timeout psutil ruff
-
-      - name: Prove MCP and telemetry trust contracts
-        env:
-          ENTROLY_DISABLE_UPDATE_CHECK: '1'
-          PYTHONUTF8: '1'
-        run: |
-          python -m compileall -q entroly tests
-          ruff check \
-            entroly/server.py \
-            entroly/value_tracker.py \
-            tests/test_value_tracker_dogfood.py \
-            tests/test_mcp_entrypoint_dogfood.py
-          python -m pytest -q --tb=short --timeout=150 \
-            tests/test_value_tracker_dogfood.py \
-            tests/test_mcp_entrypoint_dogfood.py \
-            tests/test_mcp_protocol.py
-
-      - name: Verify scope and commit product repairs
-        run: |
-          set -euo pipefail
-          test ! -e scripts/apply_final_dogfood_repairs.py
-          test ! -e .github/workflows/apply-final-dogfood-repairs.yml
-          git diff --check
-          git status --short
-          git config user.name juyterman1000
-          git config user.email 208309368+juyterman1000@users.noreply.github.com
-          git add -A
-          git commit -m 'fix(dogfood): normalize MCP counts and bound telemetry locks'
-          git push origin HEAD:dogfood/deep-runtime-20260727
